Decongest index.js and modify headline section
