+++ projects_data.js
@@ -1,5 +1,26 @@
 const projectData = [
   {
+    name: 'Medimate',
+    description: "A React on Rails app used for reserving doctors' appointments. Users can also add and delete their own doctors",
+    'Featured image': {
+      mobile: {
+        src: 'images/doc2.png',
+        alt: '',
+      },
+      desktop: {
+        src: 'images/doc2.png',
+        alt: '',
+      },
+    },
+    technologies: {
+      tech1: 'Rails',
+      tech2: 'React',
+      tech3: 'PostgreSQL',
+    },
+    'Link to live version': '#',
+    'Link to source': 'https://github.com/mailsg/medimate-frontend',
+  },
+  {
     name: 'To-do list',
     description: 'This To Do list is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete when you are done. It also allows you to edit and delete a tasks as you want. It was built using Webpack, HTML, CSS',
     'Featured image': {
@@ -41,27 +62,6 @@ const projectData = [
     },
     'Link to live version': 'https://forloop94.github.io/Conference_Page/',
     'Link to source': 'https://github.com/forLoop94/Conference_Page',
-  },
-  {
-    name: 'Project Three',
-    description: 'You need to update the main page so that the projects section is created dynamically using the information stored in that JavaScript object. Remember, all of the HTML in that section is created when the page loads.',
-    'Featured image': {
-      mobile: {
-        src: 'images/Img-Placeholder(1).png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Img-Placeholder(4).png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': '#',
-    'Link to source': '#',
   },
   {
     name: 'Project Four',
