+++ styles.css
@@ -1,7 +1,6 @@
 :root {
-  --blue: #3778c2;
-  --yellow: #F4E900;
-  --deep_ash:
+  --blue: #002244;
+  --orange: #ff6b00;
 }
 
 * {
@@ -25,18 +24,17 @@ header {
   position: fixed;
   z-index: 1;
   top: 0;
-  height: 48px;
+  height: 50px;
   width: 375px;
-  /* background: #3c3a39; */
-  background: rgb(255, 255, 255, 0.6);
+  background: rgb(0, 34, 68, 0.8);
   justify-content: space-between;
+  padding: 0.5rem 1rem;
 }
 
 #menu-btn {
   width: 16px;
   top: 0;
   left: 4px;
-  margin: 26px 24px 24.36px 0;
 }
 
 .logo {
@@ -48,8 +46,34 @@ header {
   line-height: 24px;
   text-align: center;
   color: #fff;
-  margin: 18px 0 0 24px;
   text-decoration: none;
+}
+
+.logo p {
+  background-color: white;
+  color: #002244;
+  letter-spacing: 2px;
+  padding: 2px;
+  margin: 0;
+  font-size: 8px;
+  border-radius: 5px;
+  font-family: 'Arial', sans-serif;
+  white-space: nowrap;
+}
+
+.logo-left,
+.logo-right {
+  display: inline-block;
+  transition: transform 2s;
+}
+
+.logo-right {
+  background-color: #ff6b00;
+  padding: 2px;
+  color: white;
+  border-radius: 5px;
+  font-size: 14px;
+  animation: slideNames 20s linear infinite reverse;
 }
 
 .menu-bar {
@@ -68,8 +92,7 @@ header {
   height: 590px;
   width: 375px;
   padding: 120px 24px 120px 24px;
-  /* background: #1c1a19; */
-  background: linear-gradient(90deg, white, rgb(91, 88, 88));
+  background: #002244;
 }
 
 h2 {
@@ -162,7 +185,7 @@ h2 {
 }
 
 .title {
-  margin: 20px 24px 0 24px;
+  margin: 30px 24px 0 24px;
   width: 327px;
   height: 83px;
   font-family: 'Crete Round', sans-serif;
@@ -171,7 +194,7 @@ h2 {
   font-size: 40px;
   line-height: 52px;
   align-items: center;
-  color: #1c1a19;
+  color: #002244;
 }
 
 .crossline {
@@ -264,6 +287,7 @@ h2 {
   grid-template-columns: auto;
   grid-template-rows: repeat(6, auto);
   row-gap: 24px;
+  margin-top: 2rem;
 }
 
 .grid-item {
@@ -311,6 +335,7 @@ h2 {
 .project-cards,
 .project-cards-mobile {
   width: 100%;
+  height: 400px;
 }
 
 .btn-type-b {
@@ -347,7 +372,7 @@ h2 {
   position: relative;
   width: 375px;
   height: 1810px;
-  background-color: #1c1a19;
+  background-color: #002244;
   flex-direction: column;
   padding: 120px 24px 24px 24px;
 }
@@ -457,6 +482,10 @@ h2 {
   margin-top: 10px;
 }
 
+.resume-link {
+  text-decoration: none;
+}
+
 hr {
   display: none;
 }
@@ -552,7 +581,8 @@ form {
 .grid-container-desktop,
 .divider-desktop,
 .desktop-healine-image,
-.desktop-nav {
+.desktop-nav,
+.headline-right {
   display: none;
 }
 
@@ -681,6 +711,18 @@ form {
   color: red;
 }
 
+@keyframes slideNames {
+  0% {
+      transform: translateX(0);
+  }
+  50% {
+      transform: translateX(-60px); /* Adjust the distance as needed */
+  }
+  100% {
+      transform: translateX(0);
+  }
+}
+
 @media screen and (min-width: 768px) {
   main {
     width: 100%;
@@ -690,6 +732,8 @@ form {
   header {
     width: 100%;
     height: 72px;
+    z-index: 5;
+    padding: 0;
   }
 
   .logo {
@@ -700,7 +744,30 @@ form {
     width: 100%;
     height: 19%;
     position: relative;
-    padding: 264px 143px;
+    padding: 170px 143px;
+    display: flex;
+  }
+
+  .headline-right {
+    display: block;
+    width: 1300px;
+    position: relative;
+    border-radius: 50%;
+    background-color: rgb(112, 107, 107);
+  }
+
+  .headline-img {
+    position: absolute;
+    bottom: 0;
+    background: transparent;
+    z-index: 3;
+    width: 100%;
+    height: 660px;
+    border-radius: 45%;
+    filter: brightness(150%);
+    animation: img-load 500ms;
+    animation: img-load 1s ease-out 1s alternate forwards;
+    transform: scale(0);
   }
 
   .greetings-header {
@@ -726,8 +793,11 @@ form {
 
   a:hover,
   a:focus {
-    background-color: #f0eaea;
-    color: #080807;
+    background-color: #063d73;
+  }
+
+  a:active {
+    border-bottom: 2px solid var(--orange);
   }
 
   .demoButtons a {
@@ -736,6 +806,7 @@ form {
 
   .desktop-nav-content > li > a {
     margin-right: 32px;
+    padding: 0.2rem 0.4rem;
     text-decoration: none;
     color: white;
     font-family: 'Inter', sans-serif;
@@ -743,19 +814,12 @@ form {
     font-weight: 600;
     font-size: 15px;
     line-height: 20px;
+    border-radius: 3px;
   }
 
   .mail-icon {
     width: 20px;
     height: 20px;
-  }
-
-  .desktop-healine-image {
-    height: 90%;
-    width: 50%;
-    position: absolute;
-    top: 48px;
-    right: 0;
   }
 
   h2 {
@@ -764,10 +828,11 @@ form {
     font-weight: 400;
     font-size: 4rem;
     line-height: 82px;
+    animation: slideIntro 500ms ease-in-out forwards;
   }
 
   #introduction-text {
-    width: 50%;
+    width: 100%;
     font-weight: 500;
     font-size: 32px;
     line-height: 44px;
@@ -775,7 +840,7 @@ form {
   }
 
   #brief-description {
-    width: 30%;
+    width: 80%;
     font-weight: 400;
     font-size: 1rem;
     line-height: 30px;
@@ -784,7 +849,7 @@ form {
   #contact {
     position: absolute;
     flex-direction: column;
-    top: 331px;
+    top: 231px;
     left: 36px;
   }
 
@@ -970,6 +1035,7 @@ form {
     height: 90vh;
     top: 5%;
     left: 5%;
+    z-index: 7;
   }
 
   .popup-header {
@@ -1000,7 +1066,7 @@ form {
   .project-cards-desktop {
     display: grid;
     width: 100%;
-    height: 100%;
+    /* height: 100%; */
   }
 
   .extras {
@@ -1010,5 +1076,33 @@ form {
   .message-desk {
     margin-left: 15px;
     color: red;
+  }
+
+  @keyframes img-load {
+    0% {
+      transform: translateY(30%);
+    }
+
+    50% {
+      transform: translateY(-5%);
+    }
+
+    100% {
+      transform: translateY(0);
+    }
+  }
+
+  @keyframes slideIntro {
+    0% {
+      transform: translateX(-100%);
+    }
+
+    50% {
+      transform: translateX(5%);
+    }
+
+    100% {
+      transform: translateX(0);
+    }
   }
 }
